SecuritySupport and SecuritySupportImpl merged to one
- Now it is injectable
- Simplified initialization of related classes
